Move LooseObjectStepTests to enlistment per test case
