+++ GVFS/GVFS.FunctionalTests/Tests/EnlistmentPerTestCase/TestsWithEnlistmentPerTestCase.cs
@@ -1,12 +1,18 @@
 ﻿using GVFS.FunctionalTests.Tools;
 using NUnit.Framework;
-using System.IO;
 
 namespace GVFS.FunctionalTests.Tests.EnlistmentPerTestCase
 {
     [TestFixture]
     public abstract class TestsWithEnlistmentPerTestCase
     {
+        private readonly bool forcePerRepoObjectCache;
+
+        public TestsWithEnlistmentPerTestCase(bool forcePerRepoObjectCache = false)
+        {
+            this.forcePerRepoObjectCache = forcePerRepoObjectCache;
+        }
+
         public GVFSFunctionalTestEnlistment Enlistment
         {
             get; private set;
@@ -15,7 +21,14 @@ namespace GVFS.FunctionalTests.Tests.EnlistmentPerTestCase
         [SetUp]
         public virtual void CreateEnlistment()
         {
-            this.Enlistment = GVFSFunctionalTestEnlistment.CloneAndMount(GVFSTestConfig.PathToGVFS);
+            if (this.forcePerRepoObjectCache)
+            {
+                this.Enlistment = GVFSFunctionalTestEnlistment.CloneAndMountWithPerRepoCache(GVFSTestConfig.PathToGVFS, skipPrefetch: false);
+            }
+            else
+            {
+                this.Enlistment = GVFSFunctionalTestEnlistment.CloneAndMount(GVFSTestConfig.PathToGVFS);
+            }
         }
 
         [TearDown]
